+++ ui/src/app/job-details/common/debug-icons/debug-icons.component.ts
@@ -9,7 +9,7 @@ import {ResourceUtils} from "../../../shared/utils/resource-utils";
   templateUrl: './debug-icons.component.html',
   styleUrls: ['./debug-icons.component.css']
 })
-export class JobDebugIconsComponent implements OnInit {
+export class JobDebugIconsComponent {
   @Input() displayMessage: boolean;
   @Input() message: string;
   @Input() stdout: string;
@@ -20,20 +20,17 @@ export class JobDebugIconsComponent implements OnInit {
               private sanitizer:DomSanitizer) {
   }
 
-  ngOnInit() {
-    // mark URLs as safe
-  }
-
   getResourceUrl(url: string): SafeUrl {
-    if (!url) {
+    if (!url || !ResourceUtils.isResourceURL(url)) {
       return '';
     }
     return this.sanitizer.bypassSecurityTrustResourceUrl(ResourceUtils.getResourceBrowserURL(url, this.authService.userEmail));
   }
 
   getTaskDirectory(directory): SafeUrl {
-    if (directory) {
+    if (directory && ResourceUtils.isResourceURL(directory)) {
       return this.sanitizer.bypassSecurityTrustResourceUrl(ResourceUtils.getDirectoryBrowserURL(directory, this.authService.userEmail));
     }
+    return '';
   }
 }