+++ ui/src/app/job-details/tabs/tabs.component.ts
@@ -10,6 +10,7 @@ import {
   ViewChild
 } from '@angular/core';
 import {DataSource} from '@angular/cdk/collections';
+import {MatDialog} from "@angular/material";
 import {Observable} from 'rxjs/Observable';
 
 import {JobMetadataResponse} from '../../shared/model/JobMetadataResponse';
@@ -18,6 +19,10 @@ import {JobStatusIcon} from '../../shared/common';
 import {TaskMetadata} from '../../shared/model/TaskMetadata';
 import {JobFailuresTableComponent} from "../common/failures-table/failures-table.component";
 import {JobTimingDiagramComponent} from "./timing-diagram/timing-diagram.component";
+import {JobManagerService} from "../../core/job-manager.service";
+import {Shard} from "../../shared/model/Shard";
+import {JobScatteredAttemptsComponent} from "./scattered-attempts/scattered-attempts.component";
+import {objectNotEmpty} from '../../shared/common';
 
 @Component({
   selector: 'jm-tabs',
@@ -36,21 +41,13 @@ export class JobTabsComponent implements OnInit, OnChanges {
 
   database = new TasksDatabase(this.tasks);
   dataSource: TasksDataSource | null;
-  displayedColumns = [
-    'name',
-    'status',
-    'startTime',
-    'duration',
-    'attempts',
-    'files',
-  ];
   tabWidth: number = 1024;
+
+  constructor(private readonly jobManagerService: JobManagerService,
+              public scatteredAttemptsDialog: MatDialog) {};
 
   ngOnInit() {
     this.dataSource = new TasksDataSource(this.database);
-    if (this.hasCallCachedTask() || this.hasScatteredTask()) {
-      this.displayedColumns.splice(1, 0, "taskInfoIcons");
-    }
     if (this.tabsPanel) {
       this.tabWidth = this.tabsPanel._viewContainerRef.element.nativeElement.clientWidth;
     }
@@ -65,30 +62,16 @@ export class JobTabsComponent implements OnInit, OnChanges {
     return JobStatusIcon[status];
   }
 
-  hasCallCachedTask(): boolean {
-    if (this.tasks && this.tasks.find(t => t.callCached === true)) {
-      return true;
-    }
-    return false;
-  }
-
-  hasScatteredTask(): boolean {
-    if (this.tasks && this.tasks.find(t => t.shards !== null)) {
-      return true;
-    }
-    return false;
-  }
-
   hasFailures(): boolean {
-    return this.job.failures && (this.job.failures.length !== 0);
+    return objectNotEmpty(this.job.failures);
   }
 
-  hasInputs(): boolean {
-    return this.job.inputs && (Object.keys(this.job.inputs).length !== 0);
+  hasInputs(task:TaskMetadata): boolean {
+    return objectNotEmpty(task.inputs);
   }
 
-  hasOutputs(): boolean {
-    return this.job.outputs && (Object.keys(this.job.outputs).length !== 0);
+  hasOutputs(task:TaskMetadata): boolean {
+    return objectNotEmpty(task.outputs);
   }
 
   hasTasks(): boolean {
@@ -96,6 +79,14 @@ export class JobTabsComponent implements OnInit, OnChanges {
       let tasks: TaskMetadata[] = this.job.extensions.tasks || [];
       return tasks.length > 0;
     }
+  }
+
+  hasTaskFailures(task: TaskMetadata): boolean {
+    return objectNotEmpty(task.failureMessages);
+  }
+
+  hasOnlyOneAttempt(task: TaskMetadata): boolean {
+    return task.attempts == 1;
   }
 
   getScatteredCountTotal(task: TaskMetadata): number {
@@ -107,10 +98,10 @@ export class JobTabsComponent implements OnInit, OnChanges {
   // these are the shard statuses we care about
   getShardStatuses(): JobStatus[] {
     return [JobStatus.Succeeded,
-            JobStatus.Failed,
-            JobStatus.Running,
-            JobStatus.Submitted];
-   }
+      JobStatus.Failed,
+      JobStatus.Running,
+      JobStatus.Submitted];
+  }
 
   getShardCountByStatus(task:TaskMetadata, status:JobStatus): number {
     let result = 0
@@ -125,14 +116,65 @@ export class JobTabsComponent implements OnInit, OnChanges {
     return result;
   }
 
+  getTaskFailures(task: TaskMetadata): string {
+    if (this.hasTaskFailures(task)) {
+      return task.failureMessages.join('\n');
+    }
+  }
+
   taskIsScattered(task:TaskMetadata): boolean {
     return (task.shards && task.shards.length > 0)
+  }
+
+  populateTaskAttempts(task: TaskMetadata) {
+    this.jobManagerService.getTaskAttempts(this.job.id, this.getJobTaskName(task.name)).then((response) => {
+      task.attemptsData = response.attempts;
+    })
+  }
+
+  getJobTaskName(taskName: string): string {
+    const parts = this.job.name.split('.');
+    return parts.pop() + '.' + taskName;
   }
 
   navigateDown(id: string): void {
     if (id) {
       this.navDown.emit(id);
     }
+  }
+
+  openScatteredAttemptsDialog(task: TaskMetadata): void {
+    let trimmedShards: Shard[] = [];
+
+    // remove executionEvents, since they're not needed outside the timing diagram,
+    // and preserve start and end as Date objects for task shards
+    task.shards.forEach((shard) => {
+      let newShard: Shard = {};
+      newShard.start = new Date(shard.start);
+      newShard.end = new Date(shard.end);
+      newShard.stdout = shard.stdout;
+      newShard.stderr = shard.stderr;
+      newShard.callRoot = shard.callRoot;
+      newShard.attempts = shard.attempts;
+      newShard.shardIndex = shard.shardIndex;
+      newShard.executionStatus = shard.executionStatus;
+      newShard.failureMessages = shard.failureMessages;
+      newShard.jobId = shard.jobId;
+      trimmedShards.push(newShard);
+    });
+
+    const data = {
+      taskId: this.job.id,
+      taskName: this.getJobTaskName(task.name),
+      shards: trimmedShards
+    };
+
+    this.scatteredAttemptsDialog.open(JobScatteredAttemptsComponent, {
+      disableClose: false,
+      data: {
+        shardsData: data
+      }
+    });
   }
 }